transfered utility functions from modifiers to utils
